+++ Sources/TuistCache/Cache/CacheStoring.swift
@@ -6,21 +6,24 @@ import TuistCore
 public protocol CacheStoring {
     /// Returns if the target with the given hash exists in the cache.
     /// - Parameters:
+    ///   - name: Target's name.
     ///   - hash: Target's hash.
     /// - Returns: An observable that returns a boolean indicating whether the target is cached.
-    func exists(hash: String) -> Single<Bool>
+    func exists(name: String, hash: String) -> Single<Bool>
 
     /// For the target with the given hash, it fetches it from the cache and returns a path
     /// pointint to the .xcframework that represents it.
     ///
     /// - Parameters:
+    ///   - name: Target's name.
     ///   - hash: Target's hash.
     /// - Returns: An observable that returns a boolean indicating whether the target is cached.
-    func fetch(hash: String) -> Single<AbsolutePath>
+    func fetch(name: String, hash: String) -> Single<AbsolutePath>
 
     /// It stores the xcframework at the given path in the cache.
     /// - Parameters:
+    ///   - name: Target's name.
     ///   - hash: Hash of the target the xcframework belongs to.
     ///   - paths: Path to the files that will be stored.
-    func store(hash: String, paths: [AbsolutePath]) -> Completable
+    func store(name: String, hash: String, paths: [AbsolutePath]) -> Completable
 }