+++ Tests/TuistCacheTests/Cache/CacheTests.swift
@@ -30,76 +30,86 @@ final class CacheTests: TuistUnitTestCase {
     }
 
     func test_exists_when_in_first_cache_does_not_check_second_and_returns_true() {
-        firstCache.existsStub = { hash in
+        firstCache.existsStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             return true
         }
-        secondCache.existsStub = { _ in
+        secondCache.existsStub = { _, _ in
             XCTFail("Second cache should not be checked if first hits")
             return false
         }
-        XCTAssertTrue(try subject.exists(hash: "1234").toBlocking().single())
+        XCTAssertTrue(try subject.exists(name: "targetName", hash: "1234").toBlocking().single())
     }
 
     func test_exists_when_in_second_cache_checks_both_and_returns_true() {
-        firstCache.existsStub = { hash in
+        firstCache.existsStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             return false
         }
-        secondCache.existsStub = { hash in
+        secondCache.existsStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             return true
         }
-        XCTAssertTrue(try subject.exists(hash: "1234").toBlocking().single())
+        XCTAssertTrue(try subject.exists(name: "targetName", hash: "1234").toBlocking().single())
     }
 
     func test_exists_when_not_in_cache_checks_both_and_returns_false() {
-        firstCache.existsStub = { hash in
+        firstCache.existsStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             return false
         }
-        secondCache.existsStub = { hash in
+        secondCache.existsStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             return false
         }
-        XCTAssertFalse(try subject.exists(hash: "1234").toBlocking().single())
+        XCTAssertFalse(try subject.exists(name: "targetName", hash: "1234").toBlocking().single())
     }
 
     func test_fetch_when_in_first_cache_does_not_check_second_and_returns_path() {
-        firstCache.fetchStub = { hash in
+        firstCache.fetchStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             return "/Absolute/Path"
         }
-        secondCache.fetchStub = { _ in
+        secondCache.fetchStub = { _, _ in
             XCTFail("Second cache should not be checked if first hits")
             throw TestError("")
         }
-        XCTAssertEqual(try subject.fetch(hash: "1234").toBlocking().single(), "/Absolute/Path")
+        XCTAssertEqual(try subject.fetch(name: "targetName", hash: "1234").toBlocking().single(), "/Absolute/Path")
     }
 
     func test_fetch_when_in_second_cache_checks_both_and_returns_path() {
-        firstCache.fetchStub = { hash in
+        firstCache.fetchStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             throw TestError("")
         }
-        secondCache.fetchStub = { hash in
+        secondCache.fetchStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             return "/Absolute/Path"
         }
-        XCTAssertEqual(try subject.fetch(hash: "1234").toBlocking().single(), "/Absolute/Path")
+        XCTAssertEqual(try subject.fetch(name: "targetName", hash: "1234").toBlocking().single(), "/Absolute/Path")
     }
 
     func test_fetch_when_not_in_cache_checks_both_and_throws() {
-        firstCache.fetchStub = { hash in
+        firstCache.fetchStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             throw TestError("")
         }
-        secondCache.fetchStub = { hash in
+        secondCache.fetchStub = { name, hash in
+            XCTAssertEqual(name, "targetName")
             XCTAssertEqual(hash, "1234")
             throw TestError("")
         }
         XCTAssertThrowsSpecific(
-            try subject.fetch(hash: "1234").toBlocking().single(),
+            try subject.fetch(name: "targetName", hash: "1234").toBlocking().single(),
             TestError("")
         )
     }